requirements.txt
Django==3.0.4
djangorestframework==3.12.2
django-extensions==3.1.1
openpyxl==3.0.7
requests
selenium==.141.0
beautifulsoup4==4.9.3
pyvirtualdisplay==2.1
requests